+++ src/main/java/model/User.java
@@ -8,15 +8,15 @@ import javax.persistence.Table;
 @Table(name="user")
 public class User {
 	@Id
-	private Long user_id;
+	private int user_id;
 	private String user_name;
 	private String user_psd;
 	private int role;
 	
-	public Long getUser_id() {
+	public int getUser_id() {
 		return user_id;
 	}
-	public void setUser_id(Long user_id) {
+	public void setUser_id(int user_id) {
 		this.user_id = user_id;
 	}
 	public String getUser_name() {
